requirements.txt
ConfigArgParse==1.2.3
h5py==2.10.0
imageio==2.9.0
json5==0.9.5
kornia==0.5.6
matplotlib==3.3.2
numpy==1.18.0
opencv-python==4.4.0.46
pandas==1.1.4
Pillow==7.2.0
pytorch3d==0.3.0
scikit-image==0.17.2
scipy==1.5.4
tensorboard==2.3.0
tensorflow-gpu==2.3.1
tensorflow-graphics-gpu==1.0.0
torch==1.7.1+cu101
torchsummary==1.5.1
torchvision==0.8.2+cu101
tqdm==4.50.0
transforms3d==0.3.1
trimesh==3.8.14